chore: update application for yml, create own folder feat: add ReactiveMongoRepository, Slf4j
