--- src/main/java/com/teste/wishlist/WishlistApplication.java
@@ -1,13 +1,0 @@
-package com.teste.wishlist;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class WishlistApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(WishlistApplication.class, args);
-	}
-
-}
