--- src/main/java/com/teste/wishlist/repository/WishlistRepository.java
@@ -1,11 +1,0 @@
-package com.teste.wishlist.repository;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-
-import com.teste.wishlist.model.WishListEntity;
-
-@Repository
-public interface WishlistRepository extends MongoRepository<WishListEntity, String> {
-    WishListEntity findByuserId(String userId);
-}
