--- src/main/resources/application.properties
@@ -1,16 +1,0 @@
-# SERVER CONFIG
-spring.application.name=wishlist
-server.servlet.context-path=/wishlist
-server.error.include-message=always
-server.error.include-binding-errors=always
-
-#MONGO CONFIG
-spring.data.mongodb.host=localhost
-spring.data.mongodb.port=27017
-spring.data.mongodb.database=wishlistdb
-spring.data.mongodb.auto-index-creation=true
-
-#/api-docs endpoint custom path
-springdoc.api-docs.path=/api-docs
-springdoc.show-actuator=true
-springdoc.packagesToScan=com.teste.wishlist.controller
