+++ wishlist/src/main/java/com/teste/wishlist/WishlistApplication.java
@@ -1,0 +1,20 @@
+package com.teste.wishlist;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.data.mongodb.repository.config.EnableReactiveMongoRepositories;
+
+import com.github.cloudyrock.spring.v5.EnableMongock;
+
+@EnableMongock
+@EnableReactiveMongoRepositories
+@EnableDiscoveryClient
+@SpringBootApplication
+public class WishlistApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(WishlistApplication.class, args);
+	}
+
+}
