+++ wishlist/src/main/java/com/teste/wishlist/config/ApplicationConfig.java
