+++ wishlist/src/main/java/com/teste/wishlist/config/exception/GlobalExceptionHandler.java
