+++ wishlist/src/main/java/com/teste/wishlist/model/ProductEntity.java
