+++ wishlist/src/main/java/com/teste/wishlist/service/interfaces/IProductService.java
@@ -1,0 +1,14 @@
+package com.teste.wishlist.service.interfaces;
+
+import com.teste.wishlist.model.ProductEntity;
+
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
+
+public interface IProductService {
+    Flux<ProductEntity> getAllProducts();
+
+    Mono<ProductEntity> createProduct(ProductEntity productEntity);
+
+    Mono<ProductEntity> getProductByEan(String ean);
+}
