+++ wishlist/src/test/java/com/teste/wishlist/config/exception/GlobalExceptionHandlerTest.java
