+++ wishlist/src/test/java/com/teste/wishlist/model/ProductEntityTest.java
