+++ wishlist/src/test/java/com/teste/wishlist/model/WishListEntityTest.java
