+++ wishlist/src/test/java/com/teste/wishlist/model/dto/ProductDtoTest.java
